+++ types/domain/usecases/DefaultSilentSessionMessageComposer.d.ts
@@ -1,5 +1,6 @@
 import { SilentSessionMessageComposer, SubAgentStallSections } from './adapter-interfaces/SilentSessionMessageComposer';
 export declare const composeOwnerCallFormatGuidance: () => string;
+export declare const composeMainStalledSelfDiagnosisGuidance: () => string;
 export declare class DefaultSilentSessionMessageComposer implements SilentSessionMessageComposer {
     composeMainStalledSection: (mainSilentSeconds: number) => string;
     composeMainStalledWithStaleOwnerCallSection: (mainSilentSeconds: number, unansweredOwnerCallAgeSeconds: number) => string;